Make multithreaded unsubscribe unsafe
I do not have a good way in mind to deal with unsubscribing when using
multithreaded LCM. Requiring function pointers (instead of closures)
would solve a lot of this, but those basically always require a global
mutable state and have the unfortunate side effect of not being as nice
to use as closures. I'll keep thinking on it, but this is probably where
it will land for some time.