Change hash values of primitives
Closes #26

It is currently possible to send primitives through the Rust LCM
interface without defining an LCM message type. I think this is rather
nifty and would like to keep that feature. However, without this change
it is possible, even likely, that the type you send through the
publisher is not the type you get on the other side. Having different
hashes for the primitives fixes that without breaking anything
(primitives are ignored in the hash generation of actual message types).